feat: Initial creation of Game Settings, moved Trail state from gameStateService to gameSettingsService, update the trail publish using gameSettingsService (FM-380).
