+++ src/gameSettingsService/index.ts
@@ -1,0 +1,4 @@
+import { GameSettingsService } from './game-settings-service';
+const gameSettingsServiceInstance = new GameSettingsService();
+
+export default gameSettingsServiceInstance;
