+++ tsconfig.json
@@ -67,7 +67,9 @@
       "@gamepuzzles/*": ["src/gamepuzzles/*"],
       "@gamepuzzles": ["src/gamepuzzles"],
       "@gameStateService/*": ["src/gameStateService/*"],
-      "@gameStateService": ["src/gameStateService"]
+      "@gameStateService": ["src/gameStateService"],
+      "@gameSettingsService/*": ["src/gameSettingsService/*"],
+      "@gameSettingsService": ["src/gameSettingsService"]
     },
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
