+++ webpack.config.js
@@ -61,7 +61,8 @@ var config = {
       '@events': path.resolve(__dirname, 'src/events/'),
       '@feedbackText': path.resolve(__dirname, 'src/components/feedback-text/'),
       '@gamepuzzles': path.resolve(__dirname, 'src/gamepuzzles/'),
-      '@gameStateService': path.resolve(__dirname, 'src/gameStateService/')
+      '@gameStateService': path.resolve(__dirname, 'src/gameStateService/'),
+      '@gameSettingsService': path.resolve(__dirname, 'src/gameSettingsService/')
     },
     extensions: ['.tsx', '.ts', '.js', '.json', '.css', '.sh', '.babelrc', '.eslintignore', '.gitignore', '.d'],
   },
